+++ kea-vscode/src/lru-cache/cache.ts
@@ -1,145 +1,8 @@
-interface ICacheNode<T> {
-  value: T;
-}
+import { InnerCache } from "./inner-cache";
+import { CacheKey, EndpointCache, ILinkedListNode, LinkedList, Method, MethodCache, RepositoryCache, UserCache } from "./linked-list";
 
-interface ICacheNodeValue<T> {
-  key: CacheKey;
-  value: T;
-  linkedListNode: ILinkedListNode;
-}
-
-type Method = "GET" | "POST";
-
-type MethodCache = ICacheNode<Map<Method, ICacheNodeValue<unknown>>>;
-type EndpointCache = ICacheNode<Map<string, MethodCache>>;
-type RepositoryCache = ICacheNode<Map<string, EndpointCache>>;
-type UserCache = ICacheNode<Map<string, RepositoryCache>>;
-
-type CacheKey = [user: string, repo: string, endpoint: string, method: Method];
-
-interface ILinkedListNode {
-  prev: ILinkedListNode | null;
-  next: ILinkedListNode | null;
-  key: CacheKey;
-}
-
-class LinkedList {
-  #head: ILinkedListNode | null = null;
-  #tail: ILinkedListNode | null = null;
-
-  add = (key: CacheKey): ILinkedListNode => {
-    const newNode: ILinkedListNode = {
-      prev: this.#tail,
-      next: null,
-      key,
-    };
-
-    if (this.#head === null) {
-      this.#head = newNode;
-    }
-
-    if (this.#tail !== null) {
-      this.#tail.next = newNode;
-    }
-    this.#tail = newNode;
-    return newNode;
-  };
-
-  pop = (): CacheKey | undefined => {
-    if (this.#tail === null) {
-      return;
-    }
-
-    const poppedNode = this.#tail;
-    if (this.#tail.prev === null) {
-      this.#head = null;
-      this.#tail = null;
-      return poppedNode.key;
-    }
-
-    this.#tail = poppedNode.prev;
-    if (this.#tail !== null) {
-      this.#tail.next = null;
-    }
-
-    return poppedNode.key;
-  };
-
-  promote = (node: ILinkedListNode): void => {
-    const parentNode = node.prev;
-    const grandParentNode = parentNode?.prev ?? null;
-    const nextNode = node.next;
-
-    // [grandParentNode, parentNode, node, nextNode] becomes
-    // [grandParentNode, node, parentNode, nextNode]
-
-    if (grandParentNode === null) {
-      this.#head = node;
-    } else {
-      grandParentNode.next = node;
-    }
-
-    if (parentNode !== null) {
-      parentNode.next = nextNode;
-    }
-
-    if (nextNode !== null) {
-      nextNode.prev = parentNode;
-    }
-  };
-}
-
-type GetCacheResult =
-  | undefined
-  | {
-      userCache: UserCache;
-      repoCache?: RepositoryCache;
-      endpointCache?: EndpointCache;
-      methodCache?: MethodCache;
-      value?: unknown;
-      linkedListNode?: ILinkedListNode;
-    };
-
-class Cache {
-  #cache = new Map<string, UserCache>();
-
-  get = (...[user, repo, endpoint, method]: CacheKey): GetCacheResult => {
-    const userCache = this.#cache.get(user);
-    if (userCache === undefined) {
-      return undefined;
-    }
-
-    const repoCache = userCache.value.get(repo);
-    if (repoCache === undefined) {
-      return { userCache };
-    }
-
-    const endpointCache = repoCache.value.get(endpoint);
-    if (endpointCache === undefined) {
-      return { userCache, repoCache };
-    }
-
-    const cacheMethod = endpointCache.value.get(method);
-    if (cacheMethod === undefined) {
-      return { userCache, repoCache, endpointCache };
-    }
-
-    return {
-      userCache: userCache,
-      repoCache: repoCache,
-      endpointCache: endpointCache,
-      methodCache: cacheMethod,
-      value: cacheMethod.value,
-    };
-  };
-
-  set = (user: string, cacheUser: UserCache): void => {
-    this.#cache.set(user, cacheUser);
-  };
-}
-
-class LRUCache {
-  readonly #cache = new Cache();
+export class LRUCache {
+  readonly #cache = new InnerCache();
   readonly #linkedList = new LinkedList();
 
   maxSize: number;
